feat: add useTasks custom hook for managing database manipulation and state in different file from page.tsx
