+++ src/modules/tasks/entities/index.ts
@@ -1,0 +1,16 @@
+export type Priority = 'High' | 'Medium' | 'Low';
+
+export interface Task {
+    id: number;
+    name: string;
+    finished: boolean;
+    description: string;
+    priority: Priority;
+}
+
+export interface Subtask {
+    id: number;
+    taskId: number;
+    name: string;
+    finished: boolean;
+}
